autoleasecalcbaas-master/AutoLeaseCalc/requirements.txt
django==1.11.9
djangorestframework
pyjwt
datetime
mysqlclient
mysql-python
django_extensions
django-cors-headers
django-mysql
pytz